Removed data lib, moved necessary data classes to train project, updated dependencies and imports
