+++ projects/sandbox/train/pyproject.toml
@@ -8,7 +8,7 @@ license = "MIT"
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.11"
-"bbhnet.data" = {path = "../../../libs/data", develop  = true}
+ml4gw = {path = "../../../ml4gw", develop = true}
 "bbhnet.logging" = {path = "../../../libs/logging", develop = true}
 "bbhnet.trainer" = {path = "../../../libs/trainer", develop = true}
