+++ projects/sandbox/train/train/transforms/__init__.py
@@ -1,2 +1,1 @@
 from .standard_scaler import StandardScalerTransform
-from .whitening import WhiteningTransform
