+++ projects/sandbox/train/train/utils.py
@@ -7,8 +7,8 @@ import h5py
 import numpy as np
 import torch
 
-from bbhnet.data.glitch_sampler import GlitchSampler
-from bbhnet.data.waveform_injection import BBHNetWaveformInjection
+from train.glitch_sampler import GlitchSampler
+from train.waveform_injection import BBHNetWaveformInjection
 from ml4gw.distributions import Cosine, LogNormal, Uniform
 
 Tensor = TypeVar("Tensor", np.ndarray, torch.Tensor)
